LLevo tiempo sin hacer un commit. Lo que subo es relativo a la maquina de estados. Ya realiza un combo con el boton izquierdo del ratón. Hay que darle en el momento justo. De momento ese momento justo es la fase final de la animación de ataque (o agacharse a recoger un objeto, que representa la animación de ataque enlazado al primero.). El combo es AtaqueLigero-AtaqueLigero-AtagueLigero. Si le das al ataque antes de tiempo esa orden es ignorada.

+++ Src/Logic/Entity/Components/PhysicController.cpp
@@ -112,7 +112,7 @@ void CPhysicController::process(CMessage *message)
 
 
 
-			switch (maux->getInt())
+			switch (maux->getShort())
 			{
 					case Logic::LogicalPosition::LOWER_RING:
 					{
